Moved the personality class
